requirements.txt
beautifulsoup4==4.4.1
Flask==0.12.2
lxml==3.5.0
Pillow==3.1.2
requests==2.18.4
gevent==1.2.1
flask_cors
gunicorn==19.7.0
flask_restful==0.3.6
pymysql
grequests
ftfy
SQLAlchemy==1.1.15
Werkzeug==0.12.2
mysql-connector==2.1.4
mysqlclient==1.3.12
flask_cache
flask_compress
flask_migrate
flask_script==2.0.6
flask_sqlalchemy
txfcm
apscheduler==2.1.2
pylibmc
pymemcache
pyfcm
boto3